% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_iduff.R
\name{atualizar_iduff}
\alias{atualizar_iduff}
\title{Atualiza tibble de alunos a partir de novo arquivo .xls do iduff}
\usage{
atualizar_iduff(df_antes, arquivo)
}
\arguments{
\item{df_antes}{tibble com a turma original (com campos do iduff, pelo menos).}

\item{arquivo}{Nome do arquivo .xls com a turma atual.}
}
\value{
Lista com 3 tibbles:
\itemize{
\item turma atual (com alunos excluídos marcados como inativos)
\item alunos novos
\item alunos excluídos
}
}
\description{
Inclui os alunos do novo arquivo que não estavam na tibble e
marca como inativos os alunos que não aparecem no novo arquivo.
}
\details{
Note que os alunos excluídos não são deletados da tibble, mas apenas marcados como inativos.

Note também que função retorna uma lista com 3 tibbles: a turma atual, os alunos novos, e os alunos excluídos.
}
\author{
fnaufel
}
