% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_moodle.R
\name{atualizar_moodle}
\alias{atualizar_moodle}
\title{.. content for \description{} (no empty lines) ..}
\usage{
atualizar_moodle(df_antes, arquivo)
}
\arguments{
\item{arquivo}{}
}
\value{

}
\description{
.. content for \details{} ..
}
\author{
Fernando Naufel
}
