% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_ma.R
\name{calcular_MA}
\alias{calcular_MA}
\title{Calcular média das atividades}
\usage{
calcular_MA(df, a_considerar = NA)
}
\arguments{
\item{df}{Tibble com turma}

\item{a_considerar}{Quantidade de atividades a considerar. Se NA, considerar todas as notas com nome começando com T0 ou com T1.}
}
\value{
Tibble com nova coluna MA com a média das atividades
}
\description{
Calcular média das atividades
}
\author{
fnaufel
}
