% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_metas.R
\name{calcular_metas}
\alias{calcular_metas}
\title{Calcular notas mínimas para alcançar referências}
\usage{
calcular_metas(
  df,
  nomes_notas,
  pesos_notas,
  peso_restante = 1 - sum(pesos_notas),
  referencias = c(VS = 40, AP = 60)
)
}
\arguments{
\item{df}{Data frame com as notas até agora}

\item{nomes_notas}{Nomes das notas a considerar}

\item{pesos_notas}{Pesos das notas a considerar}

\item{peso_restante}{Peso da nota que falta}

\item{referencias}{Valores de referência para o cálculo das metas}
}
\value{
Data frame original com colunas adicionais contendo as metas,
uma coluna para cada valor em \code{referencias}
}
\description{
Calcular notas mínimas para alcançar referências
}
