% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encurtar_nome.R
\name{encurtar_nomes}
\alias{encurtar_nomes}
\title{Encurta conteúdo da coluna \code{nome} de um df}
\usage{
encurtar_nomes(df, n = 2)
}
\arguments{
\item{df}{Uma tibble com a coluna \code{nome}}

\item{n}{Quantidade de nomes a manter em cada linha
(Prenome + n-1 sobrenomes)}
}
\value{
Tibble com a coluna \code{nome} modificada
}
\description{
Encurta conteúdo da coluna \code{nome} de um df
}
\author{
fnaufel
}
