% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fazer_join_por_id.R
\name{fazer_join_por_id}
\alias{fazer_join_por_id}
\title{Faz join de \code{df_antes} com \code{df_moodle} por \code{moodle_id}.}
\usage{
fazer_join_por_id(df_antes, df_moodle)
}
\arguments{
\item{df_antes}{tibble com turma anterior.}

\item{df_moodle}{tibble exportada pelo Moodle, processada por \link{ler_moodle}.}
}
\value{
tibble.
}
\description{
O resultado ainda tem as colunas de notas com os nomes dados por \link{ler_moodle}.
}
\author{
Fernando Naufel
}
