% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_alunos_iduff.R
\name{ler_alunos_iduff}
\alias{ler_alunos_iduff}
\title{Lê arquivo .xls produzido pelo iduff}
\usage{
ler_alunos_iduff(arquivo)
}
\arguments{
\item{arquivo}{Nome do arquivo .xls}
}
\value{
tibble com as colunas
\itemize{
\item \code{disciplina}
\item \code{código}
\item \code{turma}
\item \code{período}
\item \code{matrícula}
\item \code{cpf}
\item \code{nome}
\item \code{email}
\item \code{ativo}: todas as linhas contendo TRUE
\item \code{moodle}: todas as linhas contendo FALSE
}
}
\description{
Monta uma tibble com os dados da planilha.
}
\details{
Conteúdo da planilha:
\itemize{
\item Linhas 1 a 6 têm \code{disciplina}, \code{código}, \code{turma} e \code{período}.
\item Linha 7 tem os cabeçalhos das colunas: \code{matrícula}, \code{cpf}, \code{nome}, \code{email}.
}
}
\author{
Fernando Naufel
}
