% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_iduff.R
\name{ler_iduff}
\alias{ler_iduff}
\title{Lê arquivo .xls produzido pelo iduff}
\usage{
ler_iduff(arquivo)
}
\arguments{
\item{arquivo}{Nome do arquivo .xls.}
}
\value{
tibble com as colunas
\itemize{
\item \code{matricula}
\item \code{cpf}
\item \code{nome}
\item \code{email}
\item \code{disciplina}
\item \code{codigo}
\item \code{turma}
\item \code{ano}
\item \code{semestre}
\item \code{ativo}: todas as linhas contendo TRUE
}
}
\description{
Monta uma tibble com os dados da planilha.
}
\details{
Conteúdo da planilha:
\itemize{
\item Linhas 1 a 6 têm \code{disciplina}, \code{código}, \code{turma} e \code{período}.
\item Linha 7 tem os cabeçalhos das colunas: \code{matrícula}, \code{cpf}, \code{nome}, \code{email}.
}
}
\author{
fnaufel
}
