% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_moodle.R
\name{mudar_nomes}
\alias{mudar_nomes}
\title{Mudar nomes das colunas onde o moodle põe as notas}
\usage{
mudar_nomes(nome)
}
\arguments{
\item{nome}{Nome da coluna (vetorizado).}
}
\value{
Nome(s) da(s) coluna(s) alterado(s).
}
\description{
Mudar nomes das colunas onde o moodle põe as notas
}
