% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_notas.R
\name{tab_notas}
\alias{tab_notas}
\title{Constrói tabela de notas}
\usage{
tab_notas(
  df,
  legenda = NULL,
  incl_matr = FALSE,
  incl_nome = TRUE,
  ordem = ifelse(incl_nome, "nome", "matricula")
)
}
\arguments{
\item{df}{Data frame}

\item{legenda}{Vetor de strings para incluir no rodapé da tabela, Default: NULL}

\item{incl_matr}{Boolean: incluir números de matrícula?, Default: FALSE}

\item{incl_nome}{Boolean: incluir nomes dos alunos, Default: TRUE}

\item{ordem}{Coluna para ordenar, Default: ifelse(incl_nome, "nome", "matricula")}
}
\value{
Tabela kable
}
\description{
Constrói tabela de notas
}
\author{
fnaufel
}
