% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_iduff.R
\name{atualizar_iduff}
\alias{atualizar_iduff}
\title{Atualiza tibble de alunos a partir de novo arquivo .xls do iduff}
\usage{
atualizar_iduff(df_antes, arquivo)
}
\arguments{
\item{df_antes}{Tibble com a turma original (com campos do iduff, pelo
menos; talvez com os campos do moodle também).}

\item{arquivo}{Nome do arquivo .xls exportado pelo iduff.}
}
\value{
Lista com 3 tibbles:
\itemize{
\item \verb{$atual}: tibble atualizada
\item \verb{$incluidos}: tibble com novos alunos
\item \verb{$excluidos}: tibble com alunos excluídos
}
}
\description{
Inclui na tibble os alunos do novo arquivo que não estavam na
tibble e remove da tibble os alunos que não aparecem no novo arquivo.
}
\author{
fnaufel
}
