% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atualizar_moodle.R
\name{atualizar_moodle}
\alias{atualizar_moodle}
\title{Atualiza tibble de alunos a partir de novo arquivo do Moodle}
\usage{
atualizar_moodle(df_antes, arquivo)
}
\arguments{
\item{df_antes}{Tibble com dados dos alunos.}

\item{arquivo}{Caminho do arquivo Excel exportado pelo Moodle via
\verb{Grades > Export: Excel spreadsheet}.}
}
\value{
Se não houve conflitos no join, retorna a tibble atualizada.
Se houve conflitos, termina com erro.
}
\description{
Deve ser chamada para incluir na tibble as notas que foram
atribuídas aos alunos no Moodle.
}
\details{
\itemize{
\item Ignora o email do iduff em favor do email do Moodle. Isto porque os
alunos costumam alterar seus emails no Moodle, enquanto o iduff costuma
ter os emails de quando os alunos eram calouros.
\item Se a tibble já tem o campo \code{moodle_id}, usa este campo como chave do join
para os registros onde ele não é NA.
\item Senão, usa o campo \code{nome} como chave do join e emite aviso a respeito.
Se houver aluno com nome em só uma das fontes (tibble ou arquivo Moodle),
emite erro e cancela a operação.
}
}
\author{
fnaufel
}
